+++ CHANGELOG.md
@@ -1,13 +1,21 @@
 # Changelog
 
-## [1.1.9]
+## [2.0.1]
 
 ### Fixed
 
+- Updated coinbase wallet connection logic and hid unsupported ecosystems in connect wallet modal.
+
+## [2.0.0]
+
+### Fixed
+
+- Migrated to v2 routes.
 - Fixed chain and wallet logos.
 - Fixed Icon rounded issue in assets and chain, added new footer.
 - Fixed Native assets balances fetching.
 - Fixed X Link URL
+- Add dry run to all staking and distributor contract calls
 
 ## [1.1.8]